remove old proposer server impl
