+++ beacon-chain/rpc/validator/proposer_test.go
@@ -1,4 +1,4 @@
-package proposer
+package validator
 
 import (
 	"context"
